requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.5
aiosqlite==0.19.0
loguru==0.7.2
pydantic==2.7.1
python-dotenv==1.0.1 
pydantic-settings==2.3.1 